Started process of splitting up Virtio (WIP)
