+++ lib/devices/virtio/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "virtio"
+name = "virtio_device"
 version = "0.1.0"
 authors = ["Redleaf team <aburtsev@uci.edu>"]
 edition = "2018"
@@ -7,3 +7,9 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+libsyscalls = { path = "../../../lib/core/libsyscalls" }
+libtime = { path = "../../../lib/core/libtime" }
+console = { path = "../../../lib/core/console" }
+malloc = { path = "../../../lib/core/malloc" }
+rref = { path = "../../../lib/core/rref" }
+volatile_accessor = { path = "../../../lib/external/volatile_accessor" }
